api/skin-cancer-detection-backend/requirements.txt
# Core ML/Data Science packages
numpy>=1.19.5
pandas>=1.1.5
matplotlib>=3.3.4
seaborn>=0.11.1
scikit-learn>=0.24.1
opencv-python>=4.5.1.48
Pillow>=8.1.0
tensorflow>=2.8.0
keras>=2.8.0
efficientnet>=1.1.1

# FastAPI and web server
fastapi>=0.95.0
uvicorn>=0.15.0
pydantic[email]>=2.0.0
pydantic_settings>=2.0.0
email-validator>=2.0.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# MongoDB
motor>=3.3.0
pymongo>=4.5.0

# Configuration
python-dotenv==1.0.0

# Email
yagmail>=0.15.0

# File handling
pathlib>=1.0.1